moved sos from home and added emergency on bottom tab
